requirements.txt
# Core dependencies
flask==2.3.3
twilio==8.9.1
python-dotenv==1.0.0
requests==2.31.0

# AI/ML dependencies
openai==0.28.1
transformers==4.34.0
torch==2.0.1
scikit-learn==1.3.0
numpy==1.24.3
pandas==1.5.3

# Database
sqlalchemy==2.0.21

# Utilities
python-dateutil==2.8.2
pydantic==2.4.2
pyyaml==6.0.1

# Development
pytest==7.4.2
black==23.9.1
flake8==6.1.0